Separated classes into files
